JNM: Extracted GlobalConverter
